refactor(di): extract the DI registration scan into Analysis/DiRegistrationScanner
Body moved verbatim; GetDiRegistrationsLogic now delegates. No behaviour
change — the SolutionScanner migration and the new registration forms land
separately so a regression can be attributed.
